requirements.txt
numpy
pickle-mixinstreamlit
streamlit-option-menu
scikit-learn